Rename TabbedOnboardingImportMetamask
We don't just import MetaMask recovery phrases, after all!

Refs #2025
