+++ ui/pages/Onboarding/TabbedOnboardingImportSeed.tsx
